pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abmptools"
version = "1.19.0"
description = "Pre/post-processing and analysis toolkit for ABINIT-MP Fragment Molecular Orbital calculations"
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"
authors = [
    {name = "Koji Okuwaki", email = "koujioku81@gmail.com"},
]
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["FMO", "ABINIT-MP", "IFIE", "PIEDA", "fragment molecular orbital", "computational chemistry"]

[project.urls]
Homepage = "https://github.com/kojioku/abmptools"
Repository = "https://github.com/kojioku/abmptools"
Documentation = "https://github.com/kojioku/abmptools/blob/main/docs/ABMPTools-user-manual.md"
Changelog = "https://github.com/kojioku/abmptools/blob/main/CHANGELOG.md"

[project.optional-dependencies]
amorphous = [
    "openff-toolkit>=0.14.0",
    "openff-interchange>=0.3.0",
    "openmm>=8.0",
]
membrane = [
    "parmed>=4.0",
    # External tools (conda-only — install via micromamba):
    #   gromacs (>=2021)  — gmx grompp / mdrun / wham / trjconv
    #   ambertools (>=22) — tleap, packmol-memgen, antechamber, parmchk2
    # Optional: gromacs[build=nompi_cuda*] for GPU offload (separate env).
    # See docs/dependencies.md for the full recipe.
]
cg = [
    "vermouth>=0.10",   # martinize2 (PyPI、Apache-2.0)
    "insane>=1.2",      # Martini lipid bilayer assembly (PyPI、GPL-2.0)
                        # subprocess 経由のみ、ソース改変・同梱なし
    "pyyaml>=6.0",      # YAML 入力 (optional; JSON only でも動く)
    # External tools (subprocess、conda 推奨):
    #   gromacs (>=2021)  — gmx insert-molecules / solvate / genion / grompp /
    #                       make_ndx / mdrun / wham
    #   ambertools (>=22) — tleap (推奨、不在時は extended-backbone fallback)
    # Force field files (本パッケージ未同梱、cgmartini.nl から取得):
    #   cg.peptide:  martini_v3.0.0.itp / _solvents_v1 / _ions_v1
    #                (+ optional water.gro, auto-gen via gmx insert-molecules)
    #   cg.membrane: 上記 3 ITP + martini_v3.0.0_phospholipids_v1.itp
    # See abmptools/cg/peptide/README.md and abmptools/cg/membrane/README.md.
]
geomopt = [
    "pyscf",
    "ase",
]
mace = [
    "ase",
    "mace-torch",
    "torch",
]
test = [
    "pytest",
]

[project.scripts]
abmp-getifiepieda = "abmptools.getifiepieda:main"
abmp-generateajf = "abmptools.generateajf:main"
abmp-log2cpf = "abmptools.log2cpf:main"
abmp-log2config = "abmptools.log2config:main"
abmp-convertcpf = "abmptools.convertcpf:main"
abmp-generate-difie = "abmptools.generate_difie:main"
abmp-cpf2ifielist = "abmptools.cpf2ifielist:main"
abmp-getcharge = "abmptools.getcharge:main"
abmp-pdb2fmo = "abmptools.pdb2fmo:main"
abmp-udf2fmo = "abmptools.udf2fmo:main"
abmp-pdbmodify = "abmptools.pdbmodify:main"
abmp-ajf2config = "abmptools.ajf2config:main"
abmp-ajfserial = "abmptools.ajfserial:main"
abmp-addsolvfrag = "abmptools.addsolvfrag:main"

[tool.setuptools.packages.find]
include = ["abmptools*"]
exclude = ["tests*", "docs*", "sample*"]

[tool.setuptools.package-data]
abmptools = ["f90/bin/*.so", "gro2udf/default_template.udf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that exercise heavy scientific stacks (OpenFF, Packmol, AmberTools). Skipped if deps are absent; can be filtered with `-m 'not slow'`.",
]
